Add result service
